+++ NEWS.md
@@ -1,3 +1,19 @@
+# rpkgkit 0.0.4 (2026-06-02)
+
+## NEW FEATURES
+
+* Added `detect_lost_glue_brace`
+
+
+## MINOR IMPROVEMENTS
+
+* made `news_md_add_entry` vectorized
+
+* made `inquire_standalone` more explicit
+
+* made `use_hexsticker` operate on README.Rmd. Added `detect_lost_glue_brace`
+
+
 # rpkgkit 0.0.3 (2026-05-23)
 
 ## MINOR IMPROVEMENTS